Fix SLATE_USER_EMAIL env variable
