+++ packages/slate-tools/cli/commands/deploy.js
@@ -2,12 +2,11 @@ const argv = require('minimist')(process.argv.slice(2));
 const chalk = require('chalk');
 const {event} = require('@shopify/slate-analytics');
 const {replace, upload} = require('@shopify/slate-sync');
-const setEnvironment = require('../../tools/webpack/set-slate-env');
+
 const packageJson = require('../../package.json');
 const promptContinueIfPublishedTheme = require('../prompts/continue-if-published-theme');
 
 event('slate-tools:deploy:start', {version: packageJson.version});
-setEnvironment(argv.env);
 
 const deploy = argv.replace ? replace : upload;
