+++ packages/slate-tools/cli/index.js
@@ -1,11 +1,21 @@
 #!/usr/bin/env node
 const spawn = require('cross-spawn');
+const chalk = require('chalk');
+const argv = require('minimist')(process.argv.slice(2));
 const analytics = require('@shopify/slate-analytics');
 const SlateConfig = require('@shopify/slate-config');
+const slateEnv = require('@shopify/slate-env');
 const packageJson = require('../package.json');
 
 const script = process.argv[2];
 const args = process.argv.slice(3);
+
+try {
+  slateEnv.assign(argv.env);
+} catch (error) {
+  console.log(chalk.red(error));
+  process.exit(1);
+}
 
 let result;
